Added File Upload
Added Image upload for products instead of links,
Added order invoice,
Also tweaked order information to show total price, author,
Now product quantity is subtracted by user's purchased quantity

+++ src/controllers/admin.js
@@ -1,6 +1,6 @@
 // All logic related to admin
 const { validationResult } = require('express-validator');
-
+const fileHelper = require('../utils/file');
 const Product = require('../models/product');
 
 exports.getAddProduct = (req, res, next) => {
@@ -17,14 +17,31 @@ exports.getAddProduct = (req, res, next) => {
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
   const author = req.body.author;
-  const imgURL = req.body.imgURL;
+  const image = req.file;
   const description = req.body.description;
   const price = req.body.price;
   const quantity = req.body.quantity;
   const userId = req.user._id;
-
+  console.log(image);
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: `uBook - Add Product`,
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      product: {
+        title: title,
+        author: author,
+        description: description,
+        price: price,
+        quantity: quantity,
+        userId: userId,
+      },
+      errorMessage: 'Attached file is not an image!',
+      validationErrors: [],
+    });
+  }
   const errors = validationResult(req);
-
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
       pageTitle: `uBook - Add Product`,
@@ -34,7 +51,6 @@ exports.postAddProduct = (req, res, next) => {
       product: {
         title: title,
         author: author,
-        imgURL: imgURL,
         description: description,
         price: price,
         quantity: quantity,
@@ -44,10 +60,11 @@ exports.postAddProduct = (req, res, next) => {
       validationErrors: errors.array(),
     });
   }
+  const imgURL = '/' + image.path;
   const product = new Product({
     title: title,
     author: author,
-    imgURL: imgURL,
+    image: imgURL,
     description: description,
     price: price,
     quantity: quantity,
@@ -96,7 +113,7 @@ exports.postEditProduct = (req, res, next) => {
   const productId = req.body.productId;
   const title = req.body.title;
   const author = req.body.author;
-  const imgURL = req.body.imgURL;
+  const image = req.file;
   const description = req.body.description;
   const price = req.body.price;
   const quantity = req.body.quantity;
@@ -114,7 +131,6 @@ exports.postEditProduct = (req, res, next) => {
         _id: productId,
         title: title,
         author: author,
-        imgURL: imgURL,
         description: description,
         price: price,
         quantity: quantity,
@@ -128,14 +144,18 @@ exports.postEditProduct = (req, res, next) => {
     .then((product) => {
       product.title = title;
       product.author = author;
-      product.imgURL = imgURL;
       product.description = description;
       product.price = price;
       product.quantity = quantity;
       product.userId = userId;
-      return product.save();
+
+      if (image) {
+        fileHelper.deleteFile(product.image);
+        product.image = '/' + image.path;
+      }
+
+      return product.save().then((result) => res.redirect('/admin/products'));
     })
-    .then((result) => res.redirect('/admin/products'))
     .catch((err) => {
       const error = new Error(err);
       error.httpStatusCode = 500;
@@ -145,7 +165,14 @@ exports.postEditProduct = (req, res, next) => {
 
 exports.deleteProduct = (req, res, next) => {
   const productId = req.body.productId;
-  Product.findByIdAndRemove(productId)
+  Product.findById(productId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('Product was not found.'));
+      }
+      fileHelper.deleteFile(product.image);
+      return Product.findByIdAndRemove(productId);
+    })
     .then(() => {
       res.redirect('/admin/products');
     })
